requirements.txt
Pillow
unittest2
sphinx
